refactor blockly
